split implementation and header files
