+++ cpp/src/parquet/CMakeLists.txt
@@ -160,6 +160,7 @@ set(PARQUET_SRCS
     arrow/writer.cc
     bloom_filter.cc
     bloom_filter_reader.cc
+    column_chunker.cc
     column_reader.cc
     column_scanner.cc
     column_writer.cc
